Deploy to shinyapps (#140)
* Deploy to shinyapps

Closes #82.

* Add missing ).

* Fix typo in shinyapps GHA.

* Include IO in secret names

* IO in secret names in terminate

* Specify more packages to install.

* Use RStudio Public Package Manager

* More extra-packages

I suspect there will be additional cleanup once this works.

* Try to be more robust to make rsconnect happy.

* Use GITHUB_PAT.

To make installer happier. I really hope that hasn't been part of the problem all along...

* Use pak directly.

* Deal with packages that make rsconnect sad.

* Make package updater more robust.

Still technically only works if the from-cache package is available on CRAN (or the source will be specified later). Could also fail if somehow a pak dependency is installed weirdly.

* Use hyphens in app names.

Since these are different *pieces* of the name, I'd prefer to use hyphens for clearer separation.

* Delete the app, too.

And Re-install right away.

* Roll back to working script.

It isn't perfectly robust but it worked so that's better.

* Attempt to reintroduce caching.

* Caching is bad.

Also attempt to implement a thing to cancel already-running versions of these workflows. If it works, we'll institute it for other workflows as well.

* Oops. Need to install pak.