Use an app file for shinytest2.
I'm going to open an issue on shinytest2 to figure out why this isn't the recommended method, it's way cleaner.

I also pretty much emptied out the json snapshots, since those values are tested elsewhere. I really just want screenshots, but shinytest2 wants to make variants of the app on different systems for that, and it seems to work fine without those variants, so this method "tricks" it.