Load domain data once
- Closes #345.

Set up a list of domain df reactives at the top level of the server, so that we can ensure that each domain only gets fetched once per update, regardless of how many tabs/modules it is used on.

I also tried to start breaking up the monster gsmApp_Server for easier readability.