Add Domain as input.
I also added mod_ActionList back to the package. This is the first example of something that is in here *purely* for use in plugins. It might make sense for me to abstract out any of these UI convenience pieces into a completely separate package (since they aren't really specific to gsm).

-Closes #361.