Allow full-screen domain tables
Wrap domain tables in a bslib::card() so they can grow to full screen.

Note: I mentioned in #399 that maybe we'd want to do more than this, but the other parts are covered by other tickets.

I also set height and width in the screenshot tests. Hopefully that and the chromote upgrade will help make them more stable. Going to try running these tests on CI again with these updates.

- Closes #399.