Export functions used in plugins.  (#403)
* Allow full-screen domain tables

Wrap domain tables in a bslib::card() so they can grow to full screen.

Note: I mentioned in #399 that maybe we'd want to do more than this, but the other parts are covered by other tickets.

I also set height and width in the screenshot tests. Hopefully that and the chromote upgrade will help make them more stable. Going to try running these tests on CI again with these updates.

- Closes #399.

* Export functions used in plugins.

Move app to inst.

I'm not 100% certain that this shinyapps deploy will work, we'll need to test that out.

Closes #402.

* Load gsm.app locally if it isn't installed.

* More efficient pkg check.

* Move app back to root for now.