Add plugin_Define() function (#408)
* Add plugin_Define() function

Export a function to use in plugin packages, where the plugin is defined without needing a YAML file.

I also removed the AE plugin. gsm.ae depends on gsm.app, so gsm.app shouldn't also depend on gsm.ae. The demo with the AE plugin will live in that repo.

Note: I intentionally don't have the user specify packages here, because, in the situation where you're defining a plugin this way, you HAVE to define the server function "in context" of the app (either by defining the function or loading it from a package), so all of the package loading and installing happens outside of this functionality.

- Closes #406
- Closes #407

* Update vignette.