Extract app title from dfGroups (#420)
* Extract app title from dfGroups

Use nickname, protocol_title, or studyid from dfGroups to automatically set the title of the app. The title can still be overridden by the user. While I was at it, I set up the deployed apps to get a title based on what type of PR they're created from.

I also revised the navbar inputs from the previous PR to be less hacky (after a lot of experimentation). It's now much easier to extend that functionality, which I'll likely do in the next PR.

- Closes #382.

* Add a test for default condition.

I wanted to make sure it doesn't throw warnings/errors when dfGroups doesn't have any of the target fields.

* Make sure the drop-downs are always on top.