Pass plugin args with `rctv_df` prefix (#439)
* Pass plugin args with `rctv_df` prefix

If a plugin function wants `rctf_dfAE` and we have `AE`, we should pass that in. We tend to name things without the prefixes in specs, so I want the flexibility to specify them both ways.

- Closes #438

* Only use arg names that are used.

* Clean up plugin definition functions. Closes #440.

* Shared docs.

* Update data to inculde multiple snapshot dates

Note: I filter *before* calculating KRIs, so bounds are more correct and some values have changed.

* Fix dependency conflict

* Provide origin for as.Date.