Fix 489: Participant Profile Error
Don't use the `subjectid` field in the `ENROLL` domain, because it conflicts with `SubjectID`. This appears to be a gsm.mapping or other gsm.* error, where case is ignored. I should probably avoid this long-term by switching everything back to `subjid` in the domain data, I'll add a separate ticket for that... but ideally `SubjectID` and `subjectid` should be treated as different when specified that way.