Fix 489: Participant Profile Error (#490)
* Fix 489: Participant Profile Error

Don't use the `subjectid` field in the `ENROLL` domain, because it conflicts with `SubjectID`. This appears to be a gsm.mapping or other gsm.* error, where case is ignored. I should probably avoid this long-term by switching everything back to `subjid` in the domain data, I'll add a separate ticket for that... but ideally `SubjectID` and `subjectid` should be treated as different when specified that way.

* Bold Group drop-down label

Group Level and Participant stay bold, but I was accidentally un-bolding Site/Country.

* Properly pass strGroupLevel to gsmViz.default.groupOverview JS.

Closes #492.

* Install shiny v1.10.0 until v1.11.1 is released.